fix: refactor and typing
