+++ src/eirb_com_helper/chat_id_giver.py
@@ -1,7 +1,6 @@
 import asyncio
 import logging
 import sys
-from os import getenv
 
 from aiogram import Bot, Dispatcher, html
 from aiogram.client.default import DefaultBotProperties
@@ -10,10 +9,13 @@ from aiogram.filters import CommandStart
 from aiogram.types import Message
 
 from dotenv_vault import load_dotenv
+
+from .env import getenv_or_throw
+
 load_dotenv()
 
 # Bot token can be obtained via https://t.me/BotFather
-TOKEN = getenv("BOT_TOKEN")
+TOKEN = getenv_or_throw("BOT_TOKEN")
 
 # All handlers should be attached to the Router (or Dispatcher)
 
@@ -30,7 +32,7 @@ async def command_start_handler(message: Message) -> None:
     # and the target chat will be passed to :ref:`aiogram.methods.send_message.SendMessage`
     # method automatically or call API method directly via
     # Bot instance: `bot.send_message(chat_id=message.chat.id, ...)`
-    await message.answer(f"Hello, you're chat id is {html.code(message.chat.id)}!")
+    await message.answer(f"Hello, you're chat id is {html.code(str(message.chat.id))}!")
 
 
 async def run_listener() -> None:
@@ -39,7 +41,6 @@ async def run_listener() -> None:
 
     # And the run events dispatching
     await dp.start_polling(bot)
-
 
 def run():
     logging.basicConfig(level=logging.INFO, stream=sys.stdout)
